feat: port `op-proposer` from Optimism fork (#78)
* move over op proposer

* wip

* wip

* forge install: forge-std

v1.9.2

* forge install: openzeppelin-contracts

v5.0.2

* forge install: optimism

v1.9.0

* forge install: sp1-contracts

v1.1.0

* wip

* forge install: sp1-contracts

v1.1.0

* wip

* forge install: sp1-contracts

v1.1.0

* add

* forge install: sp1-contracts

* wip

* add

* forge install: openzeppelin-contracts

v4.7.3

* add

* forge install: openzeppelin-contracts

v4.7.3

* create bindings

* add

* add op-proposer-go

* fix

* fix

* fix

* add

* add

* fix

* fix

* remove run docker compose

* clean up

* fix: forge fmt

* update ci

* refactor: simplify `op-proposer` logic (#79)

* wip

* add

* fix l1Client

* add

* add comments

* update

* update

* fix: precompile override

* add value

* docs

---------


* modify

* fix

* fix

* add

* fix

* add build

---------

+++ crates/op-succinct/contracts/test/helpers/JSONDecoder.sol
@@ -3,21 +3,21 @@ pragma solidity ^0.8.0;
 
 contract JSONDecoder {
     struct Config {
-        uint chainId;
+        uint256 chainId;
         address challenger;
-        uint finalizationPeriod;
-        uint l2BlockTime;
+        uint256 finalizationPeriod;
+        uint256 l2BlockTime;
         address l2OutputOracleProxy;
         string l2RollupNode;
         address owner;
         address proposer;
-        uint startingBlockNumber;
-        uint submissionInterval;
+        uint256 startingBlockNumber;
+        uint256 submissionInterval;
         address verifierGateway;
         bytes32 vkey;
     }
 
-    struct OutputAtBlock{
+    struct OutputAtBlock {
         L2BlockRef blockRef;
         bytes32 outputRoot;
         bytes32 stateRoot;
